requirements.txt
fastapi
uvicorn
transformers
torch
sentence-transformers
scikit-learn
openai-whisper
python-docx
PyPDF2
